move transform interpolation to dedicated plugin
