+++ src/physics.rs
@@ -13,7 +13,6 @@ impl Plugin for PhysicsPlugin {
         app.add_systems(FixedUpdate, (
             velocity_move_sys,
             update_collider_aabb_sys,
-            lerp_transform_sys,
         ).run_if(in_state(GameState::Game)));
         app.add_systems(PreUpdate, collider_contact_sys.run_if(in_state(GameState::Game)));
     }
@@ -82,69 +81,4 @@ fn collider_contact_sys(
 pub struct ColliderContactEvent {
     pub a: Entity,
     pub b: Entity,
-}
-
-#[derive(Debug, Component)]
-pub struct TargetTransform {
-    transform: Transform,
-    s: f32,
-    easing_curve: EasingCurve<f32>,
-
-    pub lerp_transform: bool,
-    pub lerp_rotation: bool,
-    pub lerp_scale: bool,
-}
-
-impl TargetTransform {
-    pub fn new(target: Transform, ease_function: EaseFunction) -> Self {
-        TargetTransform {
-            transform: target,
-            s: 1.1,
-            easing_curve: EasingCurve::new(0., 1., ease_function),
-            lerp_transform: true,
-            lerp_rotation: true,
-            lerp_scale: true,
-        }
-    }
-
-    pub fn update(&mut self, target: Transform) -> &Self {
-        self.transform = target;
-        self.reset()
-    }
-
-    pub fn reset(&mut self) -> &Self {
-        self.s = 0.;
-        return self
-    }
-
-    pub fn transform(&self) -> Transform {
-        self.transform
-    }
-
-    pub fn sample_s(&self) -> f32 {
-        self.easing_curve.sample_clamped(self.s)
-    }
-}
-
-fn lerp_transform_sys(
-    mut transforms_with_targets: Query<(&mut Transform, &mut TargetTransform)>
-) {
-    for (mut tf, mut target) in transforms_with_targets.iter_mut() {
-        if target.s > 1. {
-            continue;
-        }
-
-        let s = target.sample_s();
-        if target.lerp_rotation {
-            tf.rotation = tf.rotation.lerp(target.transform.rotation, s);
-        }
-        if target.lerp_transform {
-            tf.translation = tf.translation.lerp(target.transform.translation, s);
-        }
-        if target.lerp_scale {
-            tf.scale = tf.scale.lerp(target.transform.scale, s);
-        }
-
-        target.s += 0.01;
-    }
 }
