+++ src/util/mod.rs
@@ -1,4 +1,6 @@
 pub mod ron_asset_loader;
 pub mod asset_manager;
+pub mod transform_interpolation;
 
 pub use asset_manager::{ AssetHandle, EntityAssetReadyEvent, AssetManagerPlugin };
+pub use transform_interpolation::{TargetTransform, TransformInterpolationPlugin};
